+++ calico-vpp-agent/common/pubsub.go
@@ -61,10 +61,6 @@ const (
 
 	NetAddedOrUpdated CalicoVppEventType = "NetAddedOrUpdated"
 	NetDeleted        CalicoVppEventType = "NetDeleted"
-	NetsSynced        CalicoVppEventType = "NetsSynced"
-
-	IpamPoolUpdate CalicoVppEventType = "IpamPoolUpdate"
-	IpamPoolRemove CalicoVppEventType = "IpamPoolRemove"
 )
 
 var (
