+++ config/config.go
@@ -65,7 +65,7 @@ const (
 	// BaseVppSideHardwareAddress is the base hardware address of VPP side of the HostPunt
 	// tap interface. It is used to generate hardware addresses for each uplink interface.
 	BaseVppSideHardwareAddress = "02:ca:11:c0:fd:00"
-	// CniServerStateFileVersion is the version of the CNI server state file
+	// CniServerStateFileVersion is the version of the persisted CNI state file
 	// it is used to ensure compatibility when reloading data
 	CniServerStateFileVersion = 11
 	// MaxAPITagLen is the limit number of character allowed in VPP API tags
